requirements.txt
pandas==2.0.3
joblib==1.2.0
streamlit==1.31.1
Pillow==10.0.1
scikit-learn==1.4.1.post1